hackathon/requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.15.4
gTTS==2.5.1
gunicorn==22.0.0
idna==3.6
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
msgpack==1.0.8
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==2.0.0
packaging==24.1
pillow==10.3.0
pipenv==2024.0.1
platformdirs==4.2.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.8
referencing==0.35.1
requests==2.31.0
rpds-py==0.19.1
setuptools==71.1.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.3
wheel==0.43.0
whitenoise==6.7.0
